Import village data

+++ drop_index.py
@@ -3,4 +3,5 @@ from pc_client import create_pinecone_client
 
 pc = create_pinecone_client()
 
-pc.delete_index(name=index_name)
+if pc.has_index(name=index_name):
+    pc.delete_index(name=index_name)
